client fixes
