+++ lib/fireblocks/get_transaction.go
@@ -1,0 +1,39 @@
+package fireblocks
+
+import (
+	"context"
+	"net/http"
+	"path/filepath"
+
+	"github.com/omni-network/omni/lib/errors"
+)
+
+func (c Client) GetTransactionByID(ctx context.Context, transactionID string, jwtOpts JWTOpts) (*TransactionResponse, error) {
+	var res TransactionResponse
+
+	endpoint := filepath.Join(transactionEndpoint, transactionID)
+	jwtOpts.URI = endpoint
+	jwtToken, err := genJWTToken(jwtOpts)
+	if err != nil {
+		return nil, err
+	}
+
+	response, err := c.http.SendRequest(
+		ctx,
+		endpoint,
+		http.MethodGet,
+		nil,
+		c.getHeaders(jwtToken),
+		res,
+	)
+	if err != nil {
+		return nil, err
+	}
+
+	res, ok := response.(TransactionResponse)
+	if !ok {
+		return nil, errors.New("response is not a TransactionResponse")
+	}
+
+	return &res, nil
+}
